refactor(project): split project loading and workspace model
